changing it to jest test from mocha
