+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    preset: "ts-jest",
+    testEnviroment: "node",
+    setupFilesAfterEnv: ["./jest.setup.js"],
+};
