--- test/did-document-builder.js
@@ -1,8 +1,0 @@
-const { HcsDid, DidDocumentBuilder, DidDocumentJsonProperties, DidSyntax, Hashing } = require("../dist");
-const { TopicId } = require("@hashgraph/sdk");
-const { expect, assert } = require("chai");
-
-const network = "testnet";
-const DID_TOPIC_ID = TopicId.fromString("0.0.2");
-
-describe("DidDocumentBuilder", function () {});
